backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
yfinance==0.2.32
openai==1.3.5
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
prometheus-client==0.19.0
scipy==1.11.4
pytest==7.4.3
